Use yum priorities to ensure atomic package from virt-sig
The atomic package in centos-extras got the version bumped higher
than the package we rely on from the virt-sig repo. The
centos-extras atomic package has broken dependencies when combined
with the openstack repos, so it completely breaks scenario009.

This change uses priorities to ensure we get the package from the
virt-sig repo.

Change-Id: I0aa878db62e28340d019cd92769f477189886571