Enable smoke tests and some tempest scenario on standalone
The standalone job is supposed to run quickly, and in order to keep the
quality of the tests, we are updating the list of tests being executed,
to include smoke tests, and and adding the top 3
scenarios, that will test network, storage and basic scenario. The idea
here is to keep tempest run below 1 hour.

Change-Id: I374f807ca68c4fd359395a5fea14dd302b33f0ea