Explicitly set containerized_undercloud in the UC cont. upgrade job
In change Iff509f4dc09862a451ad5cf915aa7764a314c28c we plan to
introduce smart defaulting, e.g. containerized_undercloud will take a
release-aware default value instead of always defaulting to false.

However, this change seems to break featureset050 because it changes
the containerized_undercloud value in it from false to true. Set it to
false explicitly in fs050 so that we keep the original state here and
unblock the smart defaults patch.

Change-Id: Ia54eb8988225d0b45787ab89948d2c610c27257a