master/hack: Switch to CBS instead of buildlogs
container-selinux broke all of our CI jobs and the package was untagged
in CBS but not yet synced in buildlogs.

This patch switches our CI to use CBS until buildlogs will be synced and
the broken package not pulled anymore.

This is technical debt and we don't want to use CBS on long term, so
this patch will be reverted as soon as buildlogs is updated.

Change-Id: I5e6e87822446f5a0150f25cfdd5a6184fbb68b07
Related-Bug: #1804475