Prevent reproducer failure when there is no local ssh public key
When running reproducer on a remote machine there is no guarantee
that the current user has a public key pair created as it would be
expected to be able to use they key from ssh-agent.

New code looks for local key like before but also tries to load
the first ssh agent key if the local one is not found. Previously
the code would have failed later when trying to use pub_key
variable which was pointing to a non existing file.

Change-Id: I440e61d235d8597f5ac3691d8d7415de7598bd9b