Lint fixes for playbooks and roles
This normalizes the formatting of files from playbooks and roles
folders in order to allow further hardening of linting.

Fixing linting was too big to be made in a single commit as it would
involve too many files to review and could merge conflicts with
existing changes. Thus doing it in few chunks would makes it possible.

Original full change is at https://review.openstack.org/#/c/627545/ and
will be the one merging the last.

Change-Id: I73abb9ea78b8bea8d73eb95c52daacdb45f407b8