Auto-load VIRTHOST variable when defined
Instead of failing to run due to missing target hostname, quickstart.sh
is not using VIRTHOST environment variable if defined.

And informative information message is included when this happens.

If the variable is not defined it still fails.

All error/notice messages (logging) are now send to stderr instead of
stdout as POSIX recommands (previously only some were).

Change-Id: Iadf315c14c98f4ca48b96d516d3af212e6596287