Tempest plugin url shouldn't have '.git'
When the '.git' is present at the end of the url, the Ansible task that
does the cloning will create a directory with '.git' appended.

In our case, we want the directory to be:

    /home/tempest/tempest-tripleo-ui

And not:

    /home/tempest/tempest-tripleo-ui.git

Other scripts down the line depend on the '.git' to be missing in the
directory name.

Change-Id: I0afd696234290c5ae4e5a58e5575c382f6f2877a