Add ipv4 undercloud_external_network when the overcloud is ipv6
Not all envs allow for ipv6 routing. This review adds an
ipv4 option, undercloud_external_network_ipv4, to be used
for routing/masquerade when undercloud_external_network
would return an ipv6 value.

This value is only referenced in
https://review.openstack.org/#/c/642924.
When ipv6 is not used, the undercloud_external_network_ipv4
value is never used.

Change-Id: I518bd9973bfd82350f75d2d1b2f444a5fb510a82