Install selinux shim from pypi
Instead of symlinking the selinux library from the system into the
virtualenv it will install the selinux shim library from pypi which
is detecting the library location and loading it.

This approach is more flexible as it would not break on system updates.

In fact the selinux shim library can also be installed at system
level without affecting the use of selinux imports.

Change-Id: I92e0f3a5f2117021beb060627700d9d4339c172f