Fix IPA image path used for TLS supplemental node
When deploying with a supplemental node to use TLS everywhere, a
provisioning script is used to customize the IPA image that is used
to create the supplemental node.  Most of this script uses the
$VMIMGIPA variable to refer to the image path, but one spot instead
uses the {{ ssh_user }} variable to determine the image path.  This
can result in an incorrect path depending on the user who is running
quickstart.sh, which results in a failed deployment.

This patch uses the $VMIMGIPA variable consistently in the TLS VM
provisioning script.

Change-Id: I9cf8ddf568af880ff5597006685958445e8b2569