Create supplemental node volume in correct volume pool
The tls_everywhere_provisioner.sh script creates the supplementatal
node but fails to create it in the libvirt volume pool. This mean the
idempotent check (below) fails.

- name: Check if the supplemental node volume already exists
  command: >
    virsh vol-info --pool '{{ libvirt_volume_pool }}'
    '{{ supplemental_node.name }}.qcow2'

And because virt-install creates a pool of it's choosing unknown to
oooq the pool is never cleaned up, a resource leakage.

This patch uses virsh to create the supplemental volume in the
libvirt_volume_pool where all the other node volumes are located and
then uses virsh again to upload the image into that
volume. virt-install will discover the disk image is already present
in a known pool and hence will not create a new pool.

Change-Id: Ie34233b7930badf41023635064286d2cdaab0b71
Closes-Bug: 1833456
Signed-off-by: John Dennis <jdennis@redhat.com>