Add Keycloak setup to featureset039
featureset039 is primarily used when TLS everywhere, novajoin and IPA
are needed. We want to round out this functionality with support for
federated authentication by installing the Keycloak IdP on the
supplemental node. The keycloak-setup role is being added in
tripleo-quickstart-extras as a consequence those patches appear
elsewhere in a different repository.

This patch adds the necessary configuration parameters to
featureset039 to invoke the keycloak-setup role in
tripleo-quickstart-extras and adds a dependency in
quickstart-extras-requirements.txt on an Ansible role
(nkinder.keycloak) used to install Keycloak.

Partial-Bug: 1836415

Change-Id: I59eaa2dfe7b5a3a16be18ec5d3cc03f1e68a2661
Signed-off-by: John Dennis <jdennis@redhat.com>