Pass tripleo-admin user for updates/upgrades in Queens.
fs037 and fs051, use config-download when running Queens jobs. The
config-download deployment uses tripleo-admin user for the ansible
execution, however we had hardcoded the zuul user (as Rocky, Stein
and master do use config-download by default). As a consequence, to
have the Queens jobs which use these featuresets working, we need
to specify the use of the overcloud-ssh-user to tripleo-admin.

Change-Id: I475c7c805fdcaa90b9b0183135d81f10faba3b8d
Closes-Bug: #1840763