Added  module_hotfixes support in repo setup role
repo_setup role is used to yum repos on host. In order to
install a particular package especially from a particular
repo, the .repo file needs module_hotfixes[1] needs to be
enabled. It is disabled by default.

It is needed for podman third party gating to install package
from different repo.

[1.] https://dnf.readthedocs.io/en/latest/conf_ref.html#repo-options
Change-Id: If87a0624bda0fac4f0063814f06c7f31207f390b
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>