Increase featureset020 concurrency to 4
The full tempest job is timeing out. We figure out that increasing the
concurrency makes the job pass. In some tests, increase the flavor and
keep concurrency to 3 doesn't affect the time of run. Increase the
concurrency to 4 makes the job run faster, however, increase the
concurrency to 5 make the job run slower, probably because of the lack
of resources for 5 tests running in parallel. So, 4 is the magic number.

Change-Id: I0f06f78c788b5aebf597cf5d5c025e9f6339ec39