[install-deps]Install "virtualenv<20.0.1"
New release of virtualenv broke the venv creation and
started returning 'ensure_text' module not found.
Here is the related issue:
https://github.com/pypa/virtualenv/issues/1551

In order to fix that, we need to lower the version.
Removing python-virtualenv from bindep as we are
installing it from pip.

Includes pinning of molecule as newer pre-release broke the job
while at gate.

Closes-Bug: #1862724

Change-Id: Iaf77a6009947f2eb767307b5434b60bdaf25783e
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>