Switch to images.r.o from images-vexxhost
Previously the images were hosted on rdo-cloud which was pointed
to certain IP address for rcm server and other images were hosted
on images.rdoproject.org.

In order to move the images to a new host, we can have created
images-vexxhost.r.o and tested the migration there. Since it
is working fine.

In order to reuse the name images.r.o in the new cloud, we need
to switch off the old server and populate the new dns images.r.o
on new cloud via taking an intermediate dns images-vexxhost.r.o
dns. We tested the intermediate dns.

https://softwarefactory-project.io/r/#/c/17428 populates the dns
record to images.rdoproject.org and also keeps
images-vexxhost.rdoproject.org domain. In order to keep the name
consistent, it switches to images.rdoproject.org and reuse the same.

https://tree.taiga.io/project/tripleo-ci-board/task/1538

Depends-On: https://softwarefactory-project.io/r/#/c/17428

Change-Id: Ia838e650bbab570cf2dfff4f2301497dbbc48add
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>