Include notest tag to default quickstart tags
notest tag is used in ci to skip tasks during molecule
testing, but these tasks are needed during actual deployment,
Since the default tag list do not include 'notest' tag these
tasks are skipped. This patch include that tag in default
tag list. An example of tasks which skipped due to it
were added in https://review.opendev.org/#/c/706237/ and
due to it logs are not published in ci.centos jobs.

Change-Id: Ifb4fe4acfd621de09d09859c0039344067376c92