Add native variable to locally build containers
atm there is an option to turn on local container
builds for standalone that is defined in zuul only.
ATM we are considering turning this on for all standalone
master jobs to reduce the load on docker.io

This change updates the release vars to to switch it
on for master.  By default in standalone role master
and ussuri will build locally.

Depends-On: https://review.opendev.org/#/c/743691/

Partial-Bug: #1889122
Change-Id: I32793b94525ece1e64f7924f3f29a7dc2655a588