Enable HA always for upgrade job
In Ussuri we enabled pacemaker by default so when we landed a change[0]
in quickstart to handle this logic, it broke the upgrade job because
the ussuri job uses train initially and gets a non-ha standalone which
it tries to upgrade to HA. This results in an incorrect network config.
Since really the exepectation is that we'd always be upgrade HA to HA,
let's test that instead.

[0] https://review.opendev.org/#/c/725782/
Closes-Bug: #1895822

Change-Id: I78c1a0cf68534e574b14ad505404139d93983324