Remove resource_registry args for some multinode features
This removes the resource registry args for some multinode
scenarios that are currently used. There are also a bunch of
multinode scenarios that seems to have been converted to
standalone ones. They dont even have appropriate envs in tht.
Those are not changed with this patch and should be cleaned in
a separate patch.

Change-Id: Ia40bf244f91c61b8d67eadb5b00a6cade28d00b2