Added tempest_cloud_name for fs003
Since fs003 is used in older releases like rocky and queens
for undercloud deployment where clouds.yaml is not available.

Recently we moved to openstack-tempest-skiplist where we discover
deployment type based on cloud name.

In order to enable the same, we are adding this var here to detect
the deployment.

Change-Id: I85c9f1d17d7d21a75dced66dcaa726a1409d090d
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>