Add option to add NAT networks to overcloud nodes
Jinja template for overcloud VMs is currently configured to skip
adding network to VM's definition if forward_mode: 'nat' is set
for this network. As a result, libvirt networks with enabled NAT
forwarding mode could not be used by overcloud nodes for external
connectivity.

It is not a problem for single-nic scenarios: director node is
used to connect overcloud nodes to external networks. But this
limitation makes it hard to create infrastructure for overcloud
deployments with multiple NICs and network isolation.

This patch adds flexibility by allowing users to force adding
network connections to overcloud VMs even if forward_mode: 'nat'
is set for this network. This patch doesn't change current
behavior if force_ovc parameter is not defined for any network.

Change-Id: Ibd2e40878b8fd4b1e4db7f654c251b4968da8b85
Partial-Bug: #1666916