Add octopus repos in master release file
With this patch we are updating master release file to add ceph
octopus repository in addition of ceph nautilus repository.

We haven't disabled ceph nautilus repos because ceph-ansible is not
present in Octopus repos. For now we are keeping ceph nautilius repos
as it is and just installing cephadm from octopus repo.

Once switch from ceph-ansible to cephadm is complete, We will remove
ceph nautilus repo and pull all the packages from Octopus repository.

This is with reference to patch [1] where we added Ceph octopus
repository when the release is W+.

[1] https://review.opendev.org/c/openstack/tripleo-repos/+/766683
Closes-Bug: #1909191

Change-Id: I2efd7080e7fd2922c7f18ab49ad7aca502a6c3ac