Add module_utils to ansible.cfg
Since we are using tripleo-ansible modules, we need to
include module_utils too from tripleo-ansible.

modules are already added to library path with [1]. But
since [2] module_utils are also shipped in tripleo-ansible
so we need to include those too so those can be used in
quickstart.

The issue was not seen in zuul based jobs as in these
jobs tripleo-ansible rpm is installed and tripleo-podman
module_utils got found under default module_utils path[3] which is not
true in standard quickstart based deploys where undercloud is
different node than virthost.

Also including default module path in case some other modules
need those now or in future.

[1] https://review.opendev.org/730152
[2] https://review.opendev.org/759015
[3] /usr/share/ansible/plugins/module_utils/

Change-Id: Ic9ed7a821aab77ed6a140733c397331f5851de33