Fix variable tempest_test_whitelist
The name of the variable used by os_tempest to run a specific list of
tests is tempest_test_whitelist, instead of tempest_whitelist. Usually,
we use tempest_whitelist to hold a list of tests that have conditional
to be executed, for example, if release is equal to rocky, set the
tempest_test_whitelist variable to the content of tempest_whitelist,
which is not the case here.
The undercloud-oooq job is failing because tempest is running the
default list of tests 'smoke' which have tests that are not supported to
run in the undercloud.

Change-Id: I7c264e834bd423bbe945a0c8faa4440f7e6d21f6