Remove undercloud_enable_tempest var
Since we are working in process of removing tempest container
from wallaby release and all the tempest run is switched to
os_tempest for all supported release in tripleo-ci.

This change is not going to affect any periodic jobs also.

It also removes run_tempest and tempest_format var.

Related-Bug: #1916875

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I3d046047833044fd2d75ba399f784be5d992fbef