Move all the octavia tests to fs062
octavia-tempest-plugin for train release is pinned to a
good know hash in rdoinfo [1] where smoke tests [2]does not work
and we cannot bump it to latest version as python3-httpx is not
available for train release.
It is very hard to carry overrides in the branched job due to
master parenting.

So we are moving all the tests in fs062 itself and enable the
same based on conditions.

For train periodic, we are running a good set of working tests.
Since we are already running traffic smoke tests in integration
line, so we are adding a known set of tests in check line
which will get executed in job time frame.

Links:
[1] https://review.rdoproject.org/r/c/rdoinfo/+/32936
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/32279

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Id0c32575c011e254e92578d5837fea3d1ca0a49d