Enable ephemeral_heat for featureset037
featureset037 is
tripleo-ci-centos-8-scenario000-multinode-oooq-container-updates, as
well as the -(train|ussuri/victoria) variants. However ephemeral_heat
enablement is restricted to just the wallaby+ branches in the variable
definition.

This needs to be enabled explicitly for this job before we can proceed
with https://review.opendev.org/c/789930 to use ephemeral heat by
default.

Signed-off-by: James Slagle <jslagle@redhat.com>
Depends-On: I3d8a1282c388a8201ec1ce5c7d64796c0ce6026a
Change-Id: I81d5a8e86b1b2dd03d4ea29dee10b90efe460094