Refactor inject_images for whole disk images
When whole_disk_images is true, inject_images will contain 3 items
instead of 5. This kind of conditional value cannot be set in a
variables yaml file. This change moves inject_images to a set_fact in
the undercloud setup role.

It is still possible to disable image injection by setting variable
inject_images: []. This change also includes preparation so that any
master job can be switched to using whole-disk images by setting
whole_disk_images: true. This will be set in the next change in this
series.

Change-Id: I0bd84eaa74a05d361f66f2854547f434dfb5f931
Blueprint: whole-disk-default