Suppressing ignore-errors rule in irrelevant cases
The "ignore_errors" key forces ansible to ignore failure
of the task, but not syntax errors and many other cases,
often falling under the 'error' umbrella.

One of the ansible-lint experimental rules, registers
the "ignore_errors" as a violation, and at this time it is set
to only produce warnings.

Nevertheless, while all of the cases present at this date are
a legitimate use of the "ignore_errors" mechanism,
it doesn't have to be the case in the future.

Therefore, selective suppression of the warning should give us
both information about future, problematic, violations of the rule,
and reduce the cluttering of the ansible-lint output.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Ib0558655392259c44788f514fe33d8bde202fd2f