Install tripleo-operator-ansible on cico node
As in ci.centos jobs, quickstart.sh is getting executed on
cico slave node where we donot sudo access. The playbook
execution is failing as tripleo-operator-ansible roles
are not getting discovered.

Since https://review.opendev.org/c/openstack/tripleo-quickstart/+/790701
already removed the git installation of tripleo-operator-ansible
and we cannot install it via rpm so the failure is legit.

In order to fix that, we need to install tripleo-operator-ansible
from git on cico nodes.

Since it is kept under ci-base-requirements.txt used only in
ci.centos jobs so it is not going to affect any other
deployments.

Related-Bug: #1938638

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Id7c8404914c73f3b72480ba64c0e60e9dcdf27fb