Add back collections_paths in ansible.cfg
https://review.opendev.org/c/openstack/tripleo-quickstart/+/801776
removed the collections_paths path from ansible.cfg which broke
ci.centos promotion job.

Since tripleo-operator-ansible is installed in venv in ci.centos
jobs and in order to discover that, we need the collections_path
pointing to venv in ansible.cfg.

Due to removal the job broke, adding it back to fix the issue.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Id10132bdb6f999a3384359c6cbee3b144530ed48