Make sure ovs-ofctl command exists
https://review.opendev.org/c/openstack/tripleo-quickstart/+/806955
adds Install and Restart openvswitch for upgrade to ussuri.

But it also prints version of openvswitch in the release file.
Most of the time ovs is not installed on the external node, it
leads to failure during repo-setup.

We need to make sure the command exists then print the version.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I399e08de5f69059ff5cd012b800f290eb6683936