Partial revert for tripleo-quickstart/+/791486 for centos7
This is a partial revert of [1] specifically for the centos7
case. We tested in the RDO testenv with [2] but it seems there is
some difference in the upstream which is causing related-bug.

This will allow us to unblock train while we consider how we can
use get-hash with centos7.

[1] https://review.opendev.org/c/openstack/tripleo-quickstart/+/791486
[2] https://review.rdoproject.org/r/c/testproject/+/35492

Related-Bug: 1943968
Change-Id: I010e57b9123d7694e35539619264d49d6a6cbc34