Pull tripleo-operator-ansible from current tripleo component
Currently if a change lands in tripleo-operator-ansible, we need
to wait for tripleo component promotion to get it into upstream
CI.

As our deployment is drived by tripleo-operator-ansible, it increases
the waiting time to get the fixes and delays in reviews.

In order to fix that, we need to add tripleo-operator-ansible under
required packages so that we can pull it from tripleo current repo
and test it on time.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I8514cdd9578176d867e2e1f4c104a8cb0f6811f6