Use venv instead of virtualenv on EL9.
In Centos9, virtualenv package is not available. With python3,
We don't need virtualenv and we can use "python3 -m venv"
to create virtual environments.

With this patch, Updating install-deps.sh to not install
python3-virtualenv and make use of "python3 -m venv" on
Entreprise Linux 9.

We can use use python3 -m venv" even on Centos8 as well, and even
on Centos7 as python3 is present on 7.8+. But, with this patch
we are 'only modifying what we had to' for c9 for now.

We can do a separate effort for older OS releases for cleanup of
virtualenv if we decide to get rid of virtualenv usage for older
OS releases and older distros.

Also, removing check_python_module check for virtualenv & setuptools
as this part of code does nothing even if packages are missing, and
Virtualenv/setuptools installation is already covered in prior steps.

Change-Id: I1b8bb40a4b815db6f3da1d91c816077ccb8f2ac2