Display CICO inventory before requesting node
In order to troubleshoot the fact that sometimes CICO
is not able to provide node, we need to display the inventoy
just before and after requesting a node.
If a node is freshly created but the CICO client still returns
"FATAL: no nodes were provisioned", then we'll can assume
that there is a bug in client.
If so, we'll be able to file a bug to CentOS Infra [1] with
some data.

[1] https://pagure.io/centos-infra/issues

Change-Id: I38f0f0de0d0b2b51204d1144478cb047b650f13a