Centos 9 master release file
With this patch, We are adding Centos 9 release file for master
branch. Periodic release file was added in [1]

Also, Add the bool transformation on the whole_disk_images conditional.
- On ansible the string "false" is true. The variable needs to be
transform to bool to ensure that the "false" is false and not true.
Same was done for C8 release files in [2] but was missed in c9.

[1] https://review.opendev.org/c/openstack/tripleo-quickstart/+/806910
[2] https://review.opendev.org/c/openstack/tripleo-quickstart/+/814893

Change-Id: Ibe6340017e8c8e1b5e90f783726bc364171a8ff9