Remove explicit set for whole_disk_image: true
As described in the related-bug we should stop hard coding the value
to true in the release files otherwise we cannot override it with zuul
vars.

Also, as the only role which consumes whole_disk_image, the libvirt
role now depends on extras-common.

Also explicitly pass the whole_disk_image var to the molecule test
as the extras-common role is not in play for those. Patch at [1] lets
us override with job vars or sets this according to branch.

[1] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/817243
Related-Bug: 1949765
Change-Id: I0b555676ce565d142a08e9d38d9e3c8ee7cb1a92