Use centos mirror content instead of compose
Currently CentOS stream contents are pre-released on
http://mirror.stream.centos.org/9-stream.

In TripleO Deployment, we are using compose to lay down the repos.
But on the base system, It has contents from CentOS stream mirror
where module metadata is coming and but those contents are not
available in Compose.

Due to mixing of compose and mirror content, packages
which are not available in compose are getting filtered out.

In order to avoid this issue, we are switching to
CentOS Stream mirror content. Patch at [1] adds c9 stream mirror
for opendev system-config repo.

Also removes the use of lookup plugin as it caused problems
for zuul native execution see [2]. Instead ansible_env can
give us the same value.
Also carries workaround for opendev centos9 stream repos
until [3] or similar fix merges.

[1] https://review.opendev.org/c/opendev/system-config/+/817136
[2] https://logserver.rdoproject.org/17/36817/4/check/tripleo-stream9-development-buildimage-overcloud-hardened-uefi-full-master/31ebe5e/job-output.txt
[3] https://review.opendev.org/c/opendev/base-jobs/+/820018

Closes-Bug: 1949601

Co-Authored-By: Marios Andreou (marios) <mandreou@redhat.com>
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I480dbb83d4721a30c0d96913df3e2a9d57ea4f75