Feature matrix regeneration
It appears that the generated description of oooq feature sets
was not updated in some time. This isn't an issue as long as the documentation
is build before publishing, as this regenerates the matrix.

However the version of the matrix within repository was out of sync
with the version shown in the upstream docs.[0]

Furthermore, patches to the feature sets, or their documentation,
could contain changes to the feature matrix, that were seemingly inconsistent
with the changes to the '*.yml' files.

To remedy these issues, the matrix was regenerated locally and new version
submitted as a patch.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I1b8ffeff357019ddca2232efd3ca3e1b58cd120d