Do not use whole-disk images on libvirt environments
Some quickstart development will be required for libvirt environments
around fetching/uploading the correct image and supporting
overcloud_as_undercloud when the default overcloud image is
overcloud-hardened-uefi-full.qcow2.

Now that the default on master is whole_disk_images:true, this change
sets it to false for libvirt environments so they continue to work
with overcloud-full for now.

Change-Id: I1affd48fc450250360d4eae8b30d18330a49739d
Blueprint: whole-disk-default
Closes-Bug: #1950383
Closes-Bug: #1948001