overcloud-full is uefi but VMs are defined as BIOS
Created a new parameter libvirt_boot_mode which is set depending on
the release to bios or uefi.
Depending on the libvirt_boot_mode the boot_mode is set on the
instack.json file and the libvirt xml file is set accordingly
to have whether uefi or bios boot vm.

Change-Id: Ide055d0ed8cc86af31ee512f2e4a632447750bbe
Related-Bug: 1949765