Add non collection tasks for repo-setup tripleo-get-hash
Adds a version of the repo-setup get-hash task file that is
only included when we need to avoid collections.
Using collections is problematic for zuul native plays [1].
Instead we can invoke the python cli with shell.

The default use for repo-setup remains unchanged this is
only invoked for centos9 image/container builds from [2].

[1] http://lists.zuul-ci.org/pipermail/zuul-discuss/2021-November/001752.html
[2] https://review.opendev.org/c/openstack/tripleo-ci/+/818222

Change-Id: Ic03097d92d39da3be3a85d32143256991df71fc7