Don't use ipxe-snponly on CentOS 7
Job `tripleo-ci-centos-7-containers-multinode` is failing
because packaging issue in CentOS 7.

CentOS 8 ipxe-bootimgs package has:
  /usr/share/ipxe/ipxe-i386.efi
  /usr/share/ipxe/ipxe-snponly-x86_64.efi
  /usr/share/ipxe/ipxe-x86_64.efi

CentOS 7 ipxe-bootimgs package has:
  /usr/share/ipxe/ipxe-snponly-x86_64.efi
  /usr/share/ipxe/ipxe.efi

So puppet in train does not set $arch for CentOS < 8, since the
packaging of ipxe.efi was not arch specific. But ipxe-snponly
is arch specific.

This change forces ipxe instead of ipxe-snponly for the CentOS 7
job.

Related-Bug: 1959726
Depends-On: Icfd03129bbb8645d4c6d64c8177fb9184ad8548a
Change-Id: Iffaee944c2a2bcfdad8dee1fe6c2c4a7a1d84155