Remove hardcoded undercloud baseos setting for c9
We have added discover-latest-image role(depends-on) in tqe to discover
latest Centos Stream 9 image. With this patch, Removing hardcoding
from release file for undercloud base setting for c9, We will use
the baseos info from discover-latest-image (tqe role)
to pull right base os for undercloud.

In this patch, scope is limited to use this new role in baremetal job.
We can extend the role for Centos 8 image discovery in
future but for now only using this for Centos9.(We have only
created baremetal job for C9 master for now and with c8 u/v on
eol path, c8 image discovery for train branch with no baremetal
job for train doesn't seems worth at this point.)

Depends-On: https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/840433

Change-Id: Ic214c3dd12685e85c7c794907303fdec3e70bd82