Do not fail retrieve tuned profile task
The task fails on CentOS 9-Stream nodes as no active
profile is detected due to the Related Bug[1].
So make the task not to fail til the related bug
is fixed.

The next task will set the profile based on
the value of variable "tuned_profile" if the
required profile is not already active.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2030580

Related-Bug: rhbz#2030580
Change-Id: Ib16e01540ca7769b511c6ee67b03390b8f88f3ba