Added note about 'version:' having to be a valid semantic version
The optional attribute 'version:' in entries of ansible-collection-\
requirements.yml must be a valid semantic version [1] because
ansible-galaxy will use 'version:' to compare with existing collection
installs and otherwise will fail with "Non integer values in
LooseVersion" [2].

This is in contrast to Ansible's documentation stating that 'version:'
is "The version of the role to download. Provide a tag value, commit
hash, or branch name. Defaults to master." [3].

[1] https://semver.org/
[2] https://bugzilla.redhat.com/show_bug.cgi?id=2109807
[3] https://galaxy.ansible.com/docs/using/installing.html

Change-Id: I21e30c7cef4473580848cd79f799a91c70af3bc2