Node configuration settings documentation realignment
Existing documentation of the node configuration settings was
no longer representative of the actual settings, nor of the
contents of the files referenced.

This patch changes docs to point user to correct files, and provides
more recent and accurate example of node configuration.

Code blocks were reformatted to improve readability.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I1d105501fbc56fc5c86d8381d2173129be6fe7b4