Enable auditd service in featureset010
Ansible tripleo_auditd role is now used to configure auditd service
and puppet-auditd is deprecated and archived.

Enabling auditd service in featureset010 to test auditd service in
CI.

Currently we are enabling only for master but once the backport
merges we can enable auditd service for older releases as well.

Depends-On: https://review.opendev.org/c/openstack/tripleo-heat-templates/+/852734
Change-Id: I4603db243eec097fcf454e6fa8367fb7a0c93d97