Add featureset053 for standalone roles testing
It contains the vars related for standalone ansible
roles testing.
It copies all the vars minus tempest from featureset052 to deploy
standalone in future. It contains network related vars.

It will be used for testing standalone roles in following
combination:
- Deploying compute services on the undercloud node defined
  in tripleo-ci-centos-9-singlenode-external-compute.
- Deploy the standalone alone and re-run the external data
  playbook on the same node, defined in
  tripleo-ci-centos-9-standalone-as-external-compute job. [1]

Note: Latest ansible-role-chrony is needed by tripleo-ansible
standalone roles. That's why it is added to includepkgs list
of tripleo.

[1]. https://review.opendev.org/c/openstack/tripleo-ci/+/843836

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I367f4a91b813ef1e3b768ab767de8e86f00d2ad0