Downgrade containers-common to 1-40
containers-common-1-43 adds the new keypath[1] which will
work with latest podman[2] which is not available in
podman-4.1.1-6. It breaks the deployment.

Downgrading containers-common to 1-40 fixes the issue
till we get a new podman version.

On release file changes, wallaby jobs are failing with
```
Depsolve Error occurred: \n Problem: problem with installed package catatonit-3:0.1.7-7.el9.x86_64\n
- package podman-2:4.2.0-3.el9.x86_64 conflicts with catatonit provided by catatonit-3:0.1.7-7.el9.x86_64
```
during overcloud deployment. It blocks the above changes.

We need to revert https://review.opendev.org/c/openstack/tripleo-quickstart/+/853142
the change in this patch itself and get this patch in.

Links:
[1]. https://gitlab.com/redhat/centos-stream/rpms/containers-common/-/commit/04645c4a84442da3324eea8f6538a5768e69919a
[2]. https://github.com/containers/image/commit/d218ff3d4611d35295615adf0913352a76684220

Related-Bug: #1988500
Related-Bug: #1988514
Closes-Bug: #1985981

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ie0aea674228f011881f42b9515a2e0a73198abed