Revert "[FIPS] Exclude gnutls 3.7.6 which is failing on fips"
This reverts commit 2d2fbe3ba6ca1edb6b2b31d587ef4cf6054fb0fb.

Reason for revert: this version is no longer available
Fixed version gnutls-3.7.6-11.el9 is now published:
https://bugzilla.redhat.com/show_bug.cgi?id=2119822#c7

Note: It is not a clean revert
The content provider jobs are failing with following error[1]:
```
Packages for argument containers-common-1-40.el9 available, but not installed.
```
It modifies the downgrade command to downgrade a containers-common
package only if it is installed.

[1]. https://zuul.opendev.org/t/openstack/build/75f028ea950349c9a2653f6d2aa2a157/log/logs/undercloud/home/zuul/repo_setup.log#220

Related-Bug: #1984237

Change-Id: I888b446518f30b3f63f9f566c4341a0ad289b51d