Remove conditional dlrn_hash_tag current-tripleo in mixed OS jobs
This removes the conditional used by the mixed OS jobs to pull
the right 'stable' version (in 9 line, use 8 current-tripleo)
depending on where the job is running (9 or 8 wallaby periodic).

The patch at [1] which depends-on this reworks the mixed OS
job. We will now test and qualify ci-testing for both 8 and 9
together so we need to remove this condition and always return
tripleo-ci-testing for the dlrn_hash_tag

[1] https://review.rdoproject.org/r/c/rdo-jobs/+/46119

Change-Id: Id3f25ad2990209e29e4a2792f55bafb6c6b3bdb8