Adds mixed os vars to be used with per role container image prep
This sets the vars used in [1] which depends-on this patch and
the patch at [2] will set the registry_distro_lower used here.

The repo-setup role is ammended to allow the os_version override
when resolving the centos/rhel8 tripleo-ci-testing for periodic jobs.
The override is used/passed from [1].

The release files are updated for the check/gate, integration and
component jobs.

[1] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/868210
[2] https://review.opendev.org/c/openstack/tripleo-ci/+/868207

Change-Id: If7613bd3e3bafbfe4c759b507fb6f66828a15214