Add NovaHWMachineType pin for mixed os job on rhel
The downstream rhel9/8 osp17.1 job is failing to launch a vm
during tempest tests. We need to pin the NovaHWMachineType
to something recognised (downstream only fix).
See [1] for related earlier fix.

Since we are setting extra_args here we need to remove them from
the featureset which is done in the depends-on. The extra_vars
previously set in the FS are now included here.
The vars are removed from the playbook in [2]

[1] https://review.opendev.org/c/openstack/tripleo-quickstart/+/859581
[2] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/874651

Change-Id: I54eb582d6884290c56c8fa22c70c7cea26e3796a