Pin os_tempest in tripleo jobs
We are using master branch of os_tempest from
always in all branches - train/wallaby

Recently, os-tempest dropped support for EL8
and this broke our Centos8 - train/wallaby job
and rhel8 downstream jobs.

With this patch, we are pinning os_tempest to
a hash before [1] merged.

Alternative is to use correct branch, but we
don't know how much work would that be. And
considering train will be eol soon - we are
going ahead with pinning for now.

[1] https://github.com/openstack/openstack-ansible-os_tempest/commit/2441b55aacd45cdcb8bf844852b2e293db3d6340

Closes-Bug: #2020618
Change-Id: I1568e177e6d47409ae09337d5b8775c244090513